mcp-fast/requirements.txt
fastmcp
uv
psycopg2-binary
sqlalchemy
psycopg2
python-env
py-markdown-table
pytablewriter